Moved DrawableHelper to Helpers namespace, editor hidden UI.DrawableHelper
